\name{stockAnalytics-package}
\alias{stockAnalytics-package}
\alias{stockAnalytics}
\docType{package}
\title{
\packageTitle{stockAnalytics}
}
\description{
\packageDescription{stockAnalytics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stockAnalytics}
\packageIndices{stockAnalytics}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stockAnalytics}

Maintainer: \packageMaintainer{stockAnalytics}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
